+++ tgbotapi.extensions.api/src/jvmMain/kotlin/dev/inmo/tgbotapi/extensions/api/files/DownloadFileToFile.kt
@@ -1,24 +1,29 @@
-package dev.inmo.tgbotapi.extensions.api
+package dev.inmo.tgbotapi.extensions.api.files
 
 import dev.inmo.micro_utils.coroutines.doOutsideOfCoroutine
 import dev.inmo.tgbotapi.bot.TelegramBot
 import dev.inmo.tgbotapi.extensions.api.get.getFileAdditionalInfo
-import dev.inmo.tgbotapi.requests.DownloadFile
 import dev.inmo.tgbotapi.requests.abstracts.FileId
 import dev.inmo.tgbotapi.types.files.PathedFile
 import dev.inmo.tgbotapi.types.files.abstracts.TelegramMediaFile
 import dev.inmo.tgbotapi.types.message.content.abstracts.MediaContent
+import io.ktor.util.cio.writeChannel
+import io.ktor.utils.io.copyTo
+import kotlinx.coroutines.job
 import java.io.File
+import kotlin.coroutines.coroutineContext
 
 suspend fun TelegramBot.downloadFile(
     filePath: String,
     destFile: File
 ): File {
-    val bytes = downloadFile(filePath)
-    destFile.deleteRecursively()
+    val readChannel = downloadFileStream(filePath)
 
+    destFile.deleteRecursively()
+    destFile.parentFile.mkdirs()
     doOutsideOfCoroutine { destFile.createNewFile() }
-    destFile.writeBytes(bytes)
+
+    readChannel.copyTo(destFile.writeChannel(coroutineContext.job))
 
     return destFile
 }